Add warnings for hot-reload in production and improve Dockerfile
Added a condition in the node-entrypoint.sh script to issue a warning when hot-reload is activated in a production environment. In Dockerfile, grouped the codegen and build commands into a single run instruction to optimize the creation of Docker layers. Also added comments for future improvements regarding file copying and user permissions in Dockerfile.